Add integration tests for reporters
This replaces the previous reporter tests. Instead we run the reporter
for varying test outputs and record the results. The environment is
modified to ensure the highest fidelity colors and figures are used.
Where necessary output is rewritten so the resulting logs can be
compared across Node.js versions and operating systems.

`./test/helpers/replay.js` can be used to replay a log file.

Code coverage is still pretty good, though. I'm about to refactor the
reporters so any missing coverage will be resolved at that point.