wip! Refactor IPC, status and reporters
Simplify how test results flow from the worker processes to the Api
instance to the reporters by emitting "state changes" rather than
specific events.

Workers report (most) state changes as they happen, they're collated for
each run, and reporters receive each state change and can write to the
console in turn. Stats are computed centrally in `lib/run-status.js` and
are emitted whenever they change, thus simplifying reporters.

Because workers report state changes as they happen we can greatly
simplify their coordination with the main process. We just need to make
sure the last state change has been received before exiting. If the main
process decides to exit a worker it can simplify do so, it'll already
have received results from completed tests.

Error conditions are either explicit or can be inferred from the
computed stats.

The Api now emits its plan of the tests it'll run, which can then be
sent to the reporters. This leads to better separation of concerns
between the reporters, the run status, the Api and the watch mode
implementation.

Uncaught exceptions and unhandled rejections are now shown with a code
excerpt if possible. There are other minor variations in the reporter
output, but I've tried to keep changes to a minimum.